+++ inst/sql/postgresql/migrations/Migration_4-v3_2_1_cohort_name_field.sql
@@ -1,0 +1,6 @@
+-- Set cohort name to larger field value
+{DEFAULT @migration = migration}
+{DEFAULT @cohort = cohort}
+{DEFAULT @table_prefix = ''}
+
+ALTER TABLE @database_schema.@table_prefix@cohort ALTER COLUMN cohort_name TYPE VARCHAR;