+++ inst/sql/sql_server/ComputeTimeSeries2.sql
@@ -8,7 +8,7 @@
 
 SELECT cohort_definition_id cohort_id,
 	time_id,
-	{@stratify_by_gender} ? {gender,} : {CAST(NULL AS VARCHAR) gender, }
+	{@stratify_by_gender} ? {CASE WHEN gender IS NULL THEN 'NULL' ELSE gender END} : {'NULL'} gender,
 	{@stratify_by_age_group} ? {FLOOR((YEAR(period_begin) - year_of_birth) / 10) AS age_group,} : {CAST(NULL AS INT) age_group, }
 	COUNT_BIG(DISTINCT CONCAT(cast(subject_id AS VARCHAR(30)), '_', cast(observation_period_start_date AS VARCHAR(30)))) records, -- records in calendar month
 	COUNT_BIG(DISTINCT subject_id) subjects, -- unique subjects